move dialog state to dialog control context
