+++ packages/renderer/src/contexts/import/DialogControl/index.tsx
@@ -1,0 +1,95 @@
+// Copyright 2025 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import * as defaults from './defaults';
+import { createContext, useContext, useState } from 'react';
+import type {
+  DialogBulkRenameData,
+  DialogManageAccountData,
+  DialogRenameData,
+  DialogShowAddressData,
+  DialogControlContextInterface,
+} from './types';
+
+export const DialogControlContext =
+  createContext<DialogControlContextInterface>(
+    defaults.defaultDialogControlContext
+  );
+
+export const useDialogControl = () => useContext(DialogControlContext);
+
+export const DialogControlProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  /**
+   * Rename dialog.
+   */
+  const [renameDialogState, setRenameDialogState] = useState<DialogRenameData>({
+    isOpen: false,
+    encodedAccount: null,
+    genericAccount: null,
+  });
+
+  const getRenameDialogData = () => renameDialogState;
+  const setRenameDialogData = (data: DialogRenameData) => {
+    setRenameDialogState({ ...data });
+  };
+
+  /**
+   * Manage account dialog.
+   */
+  const [manageAccountDialogState, setManageAccountDialogState] =
+    useState<DialogManageAccountData>({
+      isOpen: false,
+      genericAccount: null,
+    });
+
+  const getManageAccountDialogData = () => manageAccountDialogState;
+  const setManageAccountDialogData = (data: DialogManageAccountData) =>
+    setManageAccountDialogState({ ...data });
+
+  /**
+   * Bulk rename dialog.
+   */
+  const [bulkRenameDialogState, setBulkRenameDialogState] =
+    useState<DialogBulkRenameData>({
+      isOpen: false,
+      genericAccount: null,
+    });
+
+  const getBulkRenameDialogData = () => bulkRenameDialogState;
+  const setBulkRenameDialogData = (data: DialogBulkRenameData) =>
+    setBulkRenameDialogState({ ...data });
+
+  /**
+   * Show address dialog.
+   */
+  const [showAddressDialogState, setShowAddressDialogState] =
+    useState<DialogShowAddressData>({
+      isOpen: false,
+      address: null,
+    });
+
+  const getShowAddressDialogData = () => showAddressDialogState;
+  const setShowAddressDialogData = (data: DialogShowAddressData) =>
+    setShowAddressDialogState({ ...data });
+
+  return (
+    <DialogControlContext.Provider
+      value={{
+        getBulkRenameDialogData,
+        getManageAccountDialogData,
+        getRenameDialogData,
+        getShowAddressDialogData,
+        setBulkRenameDialogData,
+        setManageAccountDialogData,
+        setRenameDialogData,
+        setShowAddressDialogData,
+      }}
+    >
+      {children}
+    </DialogControlContext.Provider>
+  );
+};
