+++ packages/renderer/src/contexts/import/index.ts
@@ -5,6 +5,7 @@ export * from './AccountStatuses';
 export * from './AddHandler';
 export * from './Addresses';
 export * from './DeleteHandler';
+export * from './DialogControl';
 export * from './ImportHandler';
 export * from './LedgerHardware';
 export * from './RemoveHandler';
